+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Additions
 ### MLEM.Ui
 Fixes
 - Fixed auto-nav tooltip displaying on the selected element even when not in auto-nav mode
+- Fixed radio buttons not unchecking all other radio buttons with the same root element
 
 ## 5.3.0
 ### MLEM